change octave generator from float to int and increase buffer for audio engine to improve stability

+++ spectrum/src/main/java/com/lazydash/audio/visualizer/spectrum/core/algorithm/FrequencyBarsCreator.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public class FrequencyBarsCreator {
 
-    public static List<FrequencyBar> createFrequencyBars(double[] binsHz, double[] amplitudes) {
+    public static List<FrequencyBar> createFrequencyBars(int[] binsHz, double[] amplitudes) {
         List<FrequencyBar> frequencyBars = new ArrayList<>(binsHz.length);
         for (int i = 0; i < binsHz.length; i++) {
             frequencyBars.add(new FrequencyBar(binsHz[i], amplitudes[i], Color.BLACK));
