+++ spectrum/src/main/java/com/lazydash/audio/visualizer/spectrum/core/audio/procesor/FFTListener.java
@@ -1,5 +1,5 @@
 package com.lazydash.audio.visualizer.spectrum.core.audio.procesor;
 
 public interface FFTListener {
-    void frame(double[] hzBins, double[] normalizedAmplitudes);
+    void frame(int[] hzBins, double[] normalizedAmplitudes);
 }
